tests/submission/PhilipTaylor/canvas/2d.line.cap.round.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.line.cap.round</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.line.html">line</a>.<a href="index.2d.line.cap.html">cap</a>.round</h1>
<p class="desc">lineCap 'round' is rendered correctly</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.lineCap.round">2d.lineCap.round</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="green-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("lineCap 'round' is rendered correctly");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);

var tol = 1; // tolerance to avoid antialiasing artifacts

ctx.lineCap = 'round';
ctx.lineWidth = 20;


ctx.fillStyle = '#f00';
ctx.strokeStyle = '#0f0';

ctx.beginPath();
ctx.moveTo(35-tol, 15);
ctx.arc(25, 15, 10-tol, 0, Math.PI, true);
ctx.arc(25, 35, 10-tol, Math.PI, 0, true);
ctx.fill();

ctx.beginPath();
ctx.moveTo(25, 15);
ctx.lineTo(25, 35);
ctx.stroke();


ctx.fillStyle = '#0f0';
ctx.strokeStyle = '#f00';

ctx.beginPath();
ctx.moveTo(75, 15);
ctx.lineTo(75, 35);
ctx.stroke();

ctx.beginPath();
ctx.moveTo(85+tol, 15);
ctx.arc(75, 15, 10+tol, 0, Math.PI, true);
ctx.arc(75, 35, 10+tol, Math.PI, 0, true);
ctx.fill();

_assertPixel(canvas, 17,6, 0,255,0,255, "17,6", "0,255,0,255");
_assertPixel(canvas, 25,6, 0,255,0,255, "25,6", "0,255,0,255");
_assertPixel(canvas, 32,6, 0,255,0,255, "32,6", "0,255,0,255");
_assertPixel(canvas, 17,43, 0,255,0,255, "17,43", "0,255,0,255");
_assertPixel(canvas, 25,43, 0,255,0,255, "25,43", "0,255,0,255");
_assertPixel(canvas, 32,43, 0,255,0,255, "32,43", "0,255,0,255");

_assertPixel(canvas, 67,6, 0,255,0,255, "67,6", "0,255,0,255");
_assertPixel(canvas, 75,6, 0,255,0,255, "75,6", "0,255,0,255");
_assertPixel(canvas, 82,6, 0,255,0,255, "82,6", "0,255,0,255");
_assertPixel(canvas, 67,43, 0,255,0,255, "67,43", "0,255,0,255");
_assertPixel(canvas, 75,43, 0,255,0,255, "75,43", "0,255,0,255");
_assertPixel(canvas, 82,43, 0,255,0,255, "82,43", "0,255,0,255");


});
</script>

<div id="log"></div>
</body></html>
